requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12.0,<4.0.0
djangorestframework-simplejwt>=5.0.0,<6.0.0
cloudinary>=1.29.0,<2.0.0
# If you need database support
psycopg2-binary>=2.9.3,<3.0.0
# For deployment
gunicorn>=20.1.0,<21.0.0
django-cors-headers>=3.10.0,<4.0.0
# For environment variables
python-dotenv>=0.19.2,<1.0.0
cloudinary 
django-cloudinary-storage